APU Implemented
All four sound channels are working. APU has a callback field "samples_ready_func" which can be set to notify the outside environment whenever it is time to gather samples. (when sample_counter is 0)

Added "framebuffer_complete" field to the PPU which updates at the end of vblank to fix tearing. The normal framebuffer is still available.

Fixed an issue with the recent cartridges menu where the list was read backwards when the emulator starts (constantly flip flopping). Additionally, when playing a game that is already in the recent cartridges menu it will be moved to the top of the list now.